extract omniauth logic from AccountController into separate omniauth concern
